+++ kf_test.py
@@ -30,7 +30,7 @@ def KFilter_test():
     R = model.R_only_pos_meas_noise(1, sigma_v)
 
     # initial state and error convariance
-    x = np.array([1, 0.2, 2, 0.3])
+    x = np.array([1, 0.2, 2, 0.3], dtype=float)
 
     kf = ft.KFilter(F, L, H, M, Q, R, xdim, zdim)
